refactor(wallets): use shared network IDs and consolidate Stellar address validation
