+++ packages/wallets/adapters/stellar/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@layerswap/wallet-stellar",
-  "version": "2.0.0-next.0",
+  "version": "2.0.1",
   "description": "Stellar Wallet Provider for Layerswap Widget",
   "type": "module",
   "main": "dist/esm/index.js",
