+++ packages/wallets/all/tsconfig.json
@@ -13,6 +13,9 @@
   ],
   "references": [
     {
+      "path": "../../utils"
+    },
+    {
       "path": "../adapters/bitcoin"
     },
     {
